refactor: Align namespaces and minor improvement
