+++ src/Clockify/ClockifyHandler.cs
@@ -4,12 +4,8 @@ using System.Threading.Tasks;
 using Bot.Clockify.Client;
 using Bot.Clockify.Fill;
 using Bot.Clockify.Reports;
-using Bot.Dialogs;
-using Bot.Services;
-using Bot.Services.Reports;
 using Bot.States;
 using Bot.Supports;
-using Bot.Utils;
 using Luis;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
