+++ src/DIC/DicHandler.cs
@@ -1,10 +1,7 @@
 ﻿using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Bot.Clockify;
 using Bot.Dialogs;
-using Bot.Services;
-using Bot.Services.Reports;
 using Bot.States;
 using Bot.Supports;
 using Microsoft.Bot.Builder;
